refactor: improve encoder precision and enhance ListOverlay state management - Expand encoder discrete steps from uint8_t to uint16_t for greater precision - Add LVGL symbols font (Montserrat 12) to font registry - Replace CHECKED state with FOCUSED state in ListOverlay for better semantics - Add comprehensive text styling for all UI states (default, focused, pressed, disabled) - Implement recursive state application for consistent child object styling

+++ src/adapter/input/encoder/EncoderController.hpp
@@ -21,7 +21,7 @@ public:
 
     void resetEncoderPosition(EncoderID encoderId, float normalizedValue);
 
-    void setDiscreteSteps(EncoderID encoderId, uint8_t steps);
+    void setDiscreteSteps(EncoderID encoderId, uint16_t steps);
     void setContinuous(EncoderID encoderId);
 
     Encoder* getEncoder(EncoderID id);
